Added functions for primaryIdentifier
Primary Identifier is used as an alternative for displaying information about a revision's revisionable relationship model. Sometimes, ID is not enough to uniquely identify a relationship model.

Hence use primaryIdentifier functions (primaryIdentifierName, primaryIdentifierValue) with a specific column from your revisionable model to display information that allows the user to specifically identify the revision with the revisionable model instance.