Don't convert model to string unnecessarily
Using model instance in string context causes its serialization (since in Eloquent, __toString() calls toJson(), which in turn calls toArray()).
This fails if the model is not expected to be serialized when empty, or until saved.

It's enough to mention model class name for the exception message to be clear (and the exception's thrown away anyway).