Mark package as abandoned
As the number of MR's and Issues has stacked up and the original author appears to no longer wish to do anything to this (including answering people) this should really be marked as abandoned to save people wasting their time. This will flag it on packagist as abandoned, which will allow people to use alternatives.