Added the output format for values - "options".
Analogous to "boolean", only any text or numeric values can act as a source value (often flags are stored in the database). The format allows you to specify different outputs depending on the value.

options: search.On the search|network.In networks

Look at this as an associative array in which the key is separated from the value by a dot. Array elements are separated by a vertical line.